Explicitly casting init_pos to doubles

+++ doublePlasmid.c
@@ -1349,22 +1349,22 @@ void init_pos(void)
 
     if (i < nseg3 / 2)
     {
-      r3x[i] = -i - xBoxMaxd2 + nseg3;
+      r3x[i] = (double) -i - xBoxMaxd2 + nseg3;
       r3y[i] = 0.0;
     }
     if (i == nseg3 / 2 + 0.5 || i == nseg3 / 2)
     {
-      r3x[i] = -i - xBoxMaxd2 + nseg3 + 1.0;
+      r3x[i] = (double) -i - xBoxMaxd2 + nseg3 + 1.0;
       r3y[i] = -1.0;
     }
     if (i > nseg3 / 2 && i < nseg3 - 1)
     {
-      r3x[i] = +i - nseg3 - xBoxMaxd2 + nseg3 + 1.0;
+      r3x[i] = (double) +i - nseg3 - xBoxMaxd2 + nseg3 + 1.0;
       r3y[i] = -2.0;
     }
     if (i == nseg3 - 1)
     {
-      r3x[i] = +i - nseg3 - xBoxMaxd2 + nseg3 + 1.0;
+      r3x[i] = (double) +i - nseg3 - xBoxMaxd2 + nseg3 + 1.0;
       r3y[i] = -1.0;
     }
   }
@@ -1375,22 +1375,22 @@ void init_pos(void)
 
     if (i < nseg4 / 2)
     {
-      r4x[i] = +i + xBoxMaxd2 - nseg4;
+      r4x[i] = (double) +i + xBoxMaxd2 - nseg4;
       r4y[i] = 0.0;
     }
     if (i == nseg4 / 2 + 0.5 || i == nseg4 / 2)
     {
-      r4x[i] = +i + xBoxMaxd2 - nseg4 - 1.0;
+      r4x[i] = (double) +i + xBoxMaxd2 - nseg4 - 1.0;
       r4y[i] = -1.0;
     }
     if (i > nseg4 / 2 && i < nseg2 - 1)
     {
-      r4x[i] = -i - nseg4 + xBoxMaxd2 + nseg4 - 1.0;
+      r4x[i] = (double) -i - nseg4 + xBoxMaxd2 + nseg4 - 1.0;
       r4y[i] = -2.0;
     }
     if (i == nseg4 - 1)
     {
-      r4x[i] = -i - nseg4 + xBoxMaxd2 + nseg4 - 1.0;
+      r4x[i] = (double) -i - nseg4 + xBoxMaxd2 + nseg4 - 1.0;
       r4y[i] = -1.0;
     }
   }
